Fixed message for update and delete empty order and item

+++ service/static/js/rest_api.js
@@ -181,6 +181,11 @@ $(function () {
     $("#update-btn").click(function () {
 
         let order_id = $("#order_id").val();
+
+        if (!order_id || order_id == "") {
+            flash_message("Order ID is required for Update Operation")
+            return
+        }
         let customer_id = $("#order_customer_id").val();
         let status = $("#order_status").val();
 
@@ -246,6 +251,12 @@ $(function () {
         let order_id = $("#order_id").val();
 
         $("#flash_message").empty();
+
+        if (!order_id || order_id == "") {
+            clear_form_data();
+            flash_message("Order ID is required for Delete Operation")
+            return
+        }
 
         let ajax = $.ajax({
             type: "DELETE",
@@ -412,6 +423,17 @@ $(function () {
 
         let order_id = $("#order_order_id").val();
         let item_id = $("#order_item_id").val();
+
+        if (!order_id || order_id == "") {
+            flash_message("Order ID is required for Update Operation")
+            return
+        }
+
+        if (!item_id || item_id == "") {
+            flash_message("Item ID is required for Update Operation")
+            return
+        }
+
         let price = $("#order_price").val();
         let quantity = $("#order_quantity").val();
         let product_id = $("#order_product_id").val();
@@ -446,8 +468,20 @@ $(function () {
     // ****************************************
 
     $("#delete-item-btn").click(function () {
-        let order_id = $("#order_id").val();
+        let order_id = $("#order_order_id").val();
         let item_id = $("#order_item_id").val();
+
+        if (!order_id || order_id == "") {
+            clear_item_form_data();
+            flash_message("Order ID is required for Delete Operation")
+            return
+        }
+
+        if (!item_id || item_id == "") {
+            clear_item_form_data();
+            flash_message("Item ID is required for Delete Operation")
+            return
+        }
 
         $("#flash_message").empty();
